feat(Lots):
Changed from react-fontawesome to react-fa. This required updating the webpack config files to
handle the loading of the font files. I installed latest material-ui which got the icon property
working. There is some changes made to the styling of the buttons on the form. I split the question
bricks up to two groups of 3 on either side of the form.