style(NavButton):
Got a NavButton idea going. I have been grinding through the styling at this point. There are issue
since some things are inline styles an some are done using class names. The reason was to target a
material-ui paper component that is part of the dialog. Plus I need the ability to use @media query.
This looks like it can be handle with a library called
[radium](https://github.com/FormidableLabs/radium)