refactor(install): React-toolbox
Starting the switch from material-ui to react-toolbox since I would prefer to work with CSS modules
and not inline styles. I had the styles in two different location. The other solution was to go with
a library like radium or react-look to be able to handle the pseudo css class selectors like hover
and @media query. I did not want to try take on learning that at this time.