feat(UI): Contact form is back!
Got the contact form idea back working. When the user clicks on the answers div then a action is
dispatch to set it's open property to true. The Dialog component for the contact form is activated
by using the anyOpen function. Upon overlay click another action is dispatched to set the open
property on each of the question Maps to false. I have changed the open and close answer constants
to just plain open and close.