feat(Main): Changed main Attraction
After earing back from Chris I have changed the wording on the enticement messages to pain question
over general questions. I have removed the testimonials section for now and replced it with
WhatWeDo. This is the behaviour except I used recompose to make the component and a reducer jsut for
that component. This is because it is really a display component that has text as state. But instead
of using the class syntas and setState from React this keeps with the redux method and only changes
state through a reducer with dispatching action objects.