feat(Profile): Plus some setup
So the setup means setup of the other sections to be ready for this new Profile component. i have
isolated the testimonials component like the whatWeDo component. Making use of recompose to add in
its onw reducer / props. Therefore the files coupling it to the main reducers are gone. I have made
the engagment section small for mobile and just cleaned up that area a bit. I want to add a text
statement which says click for answer. This is due to the fact that on mobile there is no hover and
if the user clicks on the blue pull tab then the form opens before they see the answer to that
question.